--- a/src/Type/Definition/FieldArgument.php
+++ b/src/Type/Definition/FieldArgument.php
@@ -30,7 +30,10 @@ class FieldArgument
     /** @var mixed[] */
     public $config;
 
-    /** @var InputType&Type */
+    /**
+     * @TODO: narrow this type to Type&InputType with native asserts
+     * @var Type
+     */
     private $type;
 
     /**
@@ -40,9 +43,6 @@ class FieldArgument
     {
         foreach ($def as $key => $value) {
             switch ($key) {
-                case 'type':
-                    $this->type = $value;
-                    break;
                 case 'name':
                     $this->name = $value;
                     break;
@@ -80,7 +80,7 @@ class FieldArgument
 
     public function getType() : Type
     {
-        return Schema::resolveType($this->type);
+        return $this->type ?? ($this->type = Schema::resolveType($this->config['type']));
     }
 
     public function defaultValueExists() : bool
@@ -102,7 +102,7 @@ class FieldArgument
                 sprintf('%s.%s(%s:) %s', $parentType->name, $parentField->name, $this->name, $e->getMessage())
             );
         }
-        $type = $this->type;
+        $type = $this->getType();
         if ($type instanceof WrappingType) {
             $type = $type->getWrappedType(true);
         }