--- a/src/Type/Schema.php
+++ b/src/Type/Schema.php
@@ -373,6 +373,10 @@ class Schema
      */
     public static function resolveType($type) : Type
     {
+        if($type instanceof Type) {
+            return $type;
+        }
+
         if (is_callable($type)) {
             return $type();
         }