Hi again. I've been working with some profilers (phpbench and blackfire), and I'm afraid that while my lazy load feature does speed up my application, the performance of normal non-lazy loaded types has suffered a bit. Apologies for that, I really feel awful about it. One of the culprits is that new `Schema.resolveType` method; in my app it was getting called over 25k times (!). I came up with a few simple one-liners in this PR to lessen the pain (and drop the number of calls by an order of magnitude). Here's some profiler data to look at:

Before lazy-load-type branch:

```
\GraphQL\Benchmarks\HugeSchemaBench

    benchSchema.............................I0 [μ Mo]/r: 144.375 144.375 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSchemaLazy.........................I0 [μ Mo]/r: 0.027 0.027 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQuery.........................I0 [μ Mo]/r: 190.369 190.369 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQueryLazy.....................I0 [μ Mo]/r: 174.710 174.710 (ms) [μSD μRSD]/r: 0.000ms 0.00%

\GraphQL\Benchmarks\LexerBench

    benchIntrospectionQuery.................I4 [μ Mo]/r: 6.979 7.021 (ms) [μSD μRSD]/r: 0.069ms 0.98%

\GraphQL\Benchmarks\StarWarsBench

    benchSchema.............................I1 [μ Mo]/r: 0.522 0.522 (ms) [μSD μRSD]/r: 0.002ms 0.47%
    benchHeroQuery..........................I1 [μ Mo]/r: 2.923 2.923 (ms) [μSD μRSD]/r: 0.043ms 1.47%
    benchNestedQuery........................I1 [μ Mo]/r: 6.756 6.755 (ms) [μSD μRSD]/r: 0.026ms 0.38%
    benchQueryWithFragment..................I1 [μ Mo]/r: 7.702 7.702 (ms) [μSD μRSD]/r: 0.068ms 0.88%
    benchStarWarsIntrospectionQuery.........I1 [μ Mo]/r: 61.422 61.422 (ms) [μSD μRSD]/r: 0.205ms 0.33%

10 subjects, 19 iterations, 170 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.027 [59.578 59.583] 0.027 (ms)
⅀T: 703.026ms μSD/r 0.041ms μRSD/r: 0.452%
```

lazy-load-type branch:
```
\GraphQL\Benchmarks\HugeSchemaBench

    benchSchema.............................I0 [μ Mo]/r: 161.602 161.602 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSchemaLazy.........................I0 [μ Mo]/r: 0.033 0.033 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQuery.........................I0 [μ Mo]/r: 192.488 192.488 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQueryLazy.....................I0 [μ Mo]/r: 176.220 176.220 (ms) [μSD μRSD]/r: 0.000ms 0.00%

\GraphQL\Benchmarks\LexerBench

    benchIntrospectionQuery.................I4 [μ Mo]/r: 6.903 6.849 (ms) [μSD μRSD]/r: 0.122ms 1.76%

\GraphQL\Benchmarks\StarWarsBench

    benchSchema.............................I1 [μ Mo]/r: 0.657 0.657 (ms) [μSD μRSD]/r: 0.015ms 2.28%
    benchHeroQuery..........................I1 [μ Mo]/r: 3.012 3.012 (ms) [μSD μRSD]/r: 0.018ms 0.58%
    benchNestedQuery........................I1 [μ Mo]/r: 6.943 6.943 (ms) [μSD μRSD]/r: 0.039ms 0.57%
    benchQueryWithFragment..................I1 [μ Mo]/r: 7.727 7.727 (ms) [μSD μRSD]/r: 0.006ms 0.08%
    benchStarWarsIntrospectionQuery.........I1 [μ Mo]/r: 64.418 64.419 (ms) [μSD μRSD]/r: 0.264ms 0.41%

10 subjects, 19 iterations, 170 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.033 [62.000 61.995] 0.033 (ms)
⅀T: 730.371ms μSD/r 0.046ms μRSD/r: 0.568%
```

With these new changes:
```
\GraphQL\Benchmarks\HugeSchemaBench

    benchSchema.............................I0 [μ Mo]/r: 153.389 153.389 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSchemaLazy.........................I0 [μ Mo]/r: 0.027 0.027 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQuery.........................I0 [μ Mo]/r: 194.549 194.549 (ms) [μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQueryLazy.....................I0 [μ Mo]/r: 176.023 176.023 (ms) [μSD μRSD]/r: 0.000ms 0.00%

\GraphQL\Benchmarks\LexerBench

    benchIntrospectionQuery.................I4 [μ Mo]/r: 6.936 6.915 (ms) [μSD μRSD]/r: 0.044ms 0.63%

\GraphQL\Benchmarks\StarWarsBench

    benchSchema.............................I1 [μ Mo]/r: 0.627 0.627 (ms) [μSD μRSD]/r: 0.000ms 0.03%
    benchHeroQuery..........................I1 [μ Mo]/r: 3.172 3.172 (ms) [μSD μRSD]/r: 0.072ms 2.27%
    benchNestedQuery........................I1 [μ Mo]/r: 7.426 7.426 (ms) [μSD μRSD]/r: 0.104ms 1.40%
    benchQueryWithFragment..................I1 [μ Mo]/r: 7.934 7.934 (ms) [μSD μRSD]/r: 0.141ms 1.77%
    benchStarWarsIntrospectionQuery.........I1 [μ Mo]/r: 62.809 62.809 (ms) [μSD μRSD]/r: 0.170ms 0.27%

10 subjects, 19 iterations, 170 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.027 [61.289 61.287] 0.027 (ms)
⅀T: 722.602ms μSD/r 0.053ms μRSD/r: 0.637%
```

I also ran the same benchmark suite using blackfire:

master before lazy-load-type branch:
https://blackfire.io/profiles/fcb3e9f6-b1f8-4bb5-8fe1-adc0b5fcb5ca/graph

lazy-load-type branch:
https://blackfire.io/profiles/2249b69a-ef69-4f12-8d21-37a6b859553b/graph

this branch:
https://blackfire.io/profiles/8c31e726-a77d-4a4d-9944-2aa2ae912fe6/graph


~~So, I don't entirely understand what's happening, yet. Some of the benchmarks seem to have returned to their original values, but for some reason HugeSchemaBench.benchSchema (the one that has 600+ fields, and extracts them all) has gone up a bit. 🤔 I'll have to figure that one out in the morning.~~ edit: I think phpbench is wildly unpredictable and fluctutates a lot from run to run; it probably also doesn't help that I was running a game at the same time. I rebooted, killed other processes, and got a more stable result.

I will keep working on this until I get all the numbers back the way they were, but in the meantime I just wanted to get your thoughts and get the discussion started. Thanks as always for your suggestions. edit: I think I'm satisfied with this now, if you are. Let me know if you want me to keep working at it.