@spawnia 2020-06-29
I am going to try and do some comparison tests in my projects too.

My guess is that lazy type loading will be a big performance boost for us, since the process of actually loading a type involves constructing them from the AST representation of the SDL.

No need to wait on that though, the numbers and changes in this PR look good to me.