@shmax 2020-07-08
@mfn The thing that changed is that we now lazy load types (whether they're callable or not), so it's no longer appropriate to directly access the public `type` member on fields. I think we're going to have to remove all those public type properties, For now, you can fix your test by doing something like this:

```php
// Rules.php:142
$rules = $rules + $this->inferRulesFromType($field instanceof InputObjectField ? $field->getType() : $field->type, $key, $resolutionArguments[$name]);
```

@vladar  In the meantime, should we revert that commit and re-release it with a larger bump?