@mfn 2020-07-08
@shmax thanks, I applied your fix but there are still tests breaking

I'm tracking this in https://github.com/rebing/graphql-laravel/pull/652 , i.e. this is the commit with your suggestion change https://github.com/rebing/graphql-laravel/pull/652/commits/21c959c7c2aa2f89d82dd6ae413fcf3dd351c60a

There are multiple failing tests and I'm still tracking them, but one thing that stands out is that _within_ graphql-php there also occurs an issue with the library, see https://github.com/rebing/graphql-laravel/runs/848701923?check_suite_focus=true#step:11:29

This is this code from graphql-php https://github.com/webonyx/graphql-php/blob/6096b0174c2b40f3bff9889f3a91dcf071bae933/src/Type/Definition/FieldArgument.php#L84

Not yet sure yet what's the issue with the config though; appreciate any help.