@shmax 2020-07-08
> Not yet sure yet what's the issue with the config though; appreciate any help.
Evidently, somewhere a `FieldArgument` is being instantiated with no 'type' property passed to its constructor. I'd like to debug it, but when I try to run those tests, I get this error: 

```
Illuminate\Database\QueryException : could not find driver (SQL: select * from sqlite_master where type = 'table' and name = migrations)
```

What's the quickest way for me to get set up to run those tests? (I actually have no idea what laravel is, or what it does). I'm running in a windows environment, but I can also run in CentOS. Please advise. In the meantime, I'll open a revert PR.