@mfn 2020-07-08
@shmax I guess you need the sqlite extension for the tests, otherwise they should "just work" 🤔