@shmax 2020-07-08
Ah, gotcha. I haven't used sqllite in a long time. I'll get set up tomorrow. Thanks for hanging in there.