@shmax 2020-07-09
@mfn I spend the evening trying to figure out where this error is coming from:

```
ErrorException : Undefined index: type
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Type\Definition\FieldArgument.php:84
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Utils\TypeInfo.php:203
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Utils\TypeInfo.php:205
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Utils\TypeInfo.php:160
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Utils\TypeInfo.php:219
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Type\Schema.php:214
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Type\Schema.php:200
 D:\repos\graphql-laravel\vendor\webonyx\graphql-php\src\Type\Schema.php:153
 D:\repos\graphql-laravel\src\GraphQL.php:87
 D:\repos\graphql-laravel\src\GraphQL.php:142
 D:\repos\graphql-laravel\src\GraphQL.php:126
 D:\repos\graphql-laravel\vendor\laravel\framework\src\Illuminate\Support\Facades\Facade.php:261
 D:\repos\graphql-laravel\tests\TestCase.php:172
 D:\repos\graphql-laravel\tests\Database\SelectFields\NestedRelationLoadingTests\NestedRelationLoadingTest.php:56
```

Kind of an interesting one, really. Turns out I didn't break anything in this branch; what is happening is that a bug in your test code was being exposed. This is the troublesome bit:

```php
// UserType::29
           'posts' => [
                'type' => Type::nonNull(Type::listOf(Type::nonNull(GraphQL::type('Post')))),
                'args' => [
                    'flag' => [
                        Type::boolean(),
                    ],
                ],
```

See the problem? This is illegal syntax. Your `args` field must be comprised of a map of keys to config objects, or a map of keys to types. You have neither. You can fix the bug by doing this:

```php
           'posts' => [
                'type' => Type::nonNull(Type::listOf(Type::nonNull(GraphQL::type('Post')))),
                'args' => [
                    'flag' => [
                        'type' => Type::boolean(),
                    ],
                ],
```

Theoretically, you should also be able to do:

```php

           'posts' => [
                'type' => Type::nonNull(Type::listOf(Type::nonNull(GraphQL::type('Post')))),
                'args' => [
                    'flag' => Type::boolean()
                ],
```

But then something in `RulesInFields` breaks. I'll leave that one for you to investigate.

The mysterious part is why the code seemed to "work", before.

Basically we were passing that odd config object to our `FieldArgument` constructor, and wound up in this situation:

![image](https://user-images.githubusercontent.com/773172/86999354-3dbfe580-c167-11ea-858d-76943ffe08e8.png)

If you can immediately understand why this worked, then you are a smarter bear than me; I had to stare at it for 10 minutes.

Your malformed config object has two keys, numeric `0` and string `name`. When we walk over the keys, on our first iteration `$key` is `0`. Because it's an integer, when it compares it to the first case, it casts the string "type" to an int, which turns out to be `0. So the type configuration, by sheer accident of ordering, winds up getting assigned to `$this->type`, which is where it belongs.

Just one of those rare cases where one bug cancels another.

Meanwhile, I've opened another PR to correct the public `type` member issue on our end, but I'm not sure when that will be released, so in the meantime I thought you might want to see if you can fix as much of this stuff on your end as possible.