@mfn 2020-07-09
@shmax I lack the words to describe my embarrassment and I very much appreciate you helping me out on this one (as well as for your optimization work here 👍 )

> Because it's an integer, when PHP compares it to the first case, it casts the string "type" to an int, which turns out to be `0`.

Major PITA 💥 

I'm happy that https://wiki.php.net/rfc/match_expression_v2 passed as it would have not allowed this ambiguity in the first place 😏 

Back to https://github.com/webonyx/graphql-php/pull/684#issuecomment-655299920
> The thing that changed is that we now lazy load types (whether they're callable or not), so it's no longer appropriate to directly access the public type member on fields

To make all tests green, I identified another case where `->type` was accessed directly.

This certainly feels like a breaking change which shouldn't happen in a minor version and https://github.com/webonyx/graphql-php/pull/699 feels right (though at this point graphql-laravel is fixed so it wouldn't matter anymore and I hadn't made a release with the version bump anyway)